fix bug

+++ utilCartoon.py
@@ -44,7 +44,7 @@ api_logger.info(f"共有 {len(framePaths)} 帧")
 api_logger.info("加载模型")
 model_id = "instruction-tuning-sd/cartoonizer"
 pipeline = StableDiffusionInstructPix2PixPipeline.from_pretrained(
-    model_id, torch_dtype=torch.float32, use_auth_token=True
+    model_id, torch_dtype=torch.bfloat16, use_auth_token=True
 ).to("cuda")
 
 # image_path = "./sample/WX20240314-161847.png"
